refix: delay type changed to uint32

+++ gpioHandler.c
@@ -4,8 +4,8 @@
 
 static uint8_t ports[11] = {0,0,0,0,0,0,0,0,0,0,0};
 
-void delay(uint8_t ms){
-	for(volatile uint8_t i = 0; i < ms*MS; i++);
+void delay(uint32_t ms){
+	for(volatile uint32_t i = 0; i < ms*MS; i++);
 }
 
 void writeHigh(Pin* pin){
